chore: copied the meta data from the main testing repository over to the images repository
We need to have some of the cofiguration files locally, so we are moving all the configuration into the motra-images repository as well. we will ship the default settings and configuration files from motra-setups in the future also, but for building and publishing the images on GHCR we need to have the setup locally inside this repository. otherwise building and setup will always need to fetch sources from github:motra-setups.